workers/ocr/requirements.txt
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
minio==7.2.0
pytesseract==0.3.10
pillow==10.1.0
pymupdf==1.23.8
python-dotenv==1.0.0
requests==2.31.0
